# ayurveda-shop/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Authentication","description":"Authentication and user management endpoints"},{"name":"Products","description":"Product management endpoints"},{"name":"Orders","description":"Order management endpoints"},{"name":"Registration","description":"User registration endpoint"}],"paths":{"/api/products/{productId}":{"get":{"tags":["Products"],"summary":"Get product","description":"Get product by ID","operationId":"getProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"put":{"tags":["Products"],"summary":"Update product","description":"Update existing product","operationId":"updateProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"delete":{"tags":["Products"],"summary":"Delete product","description":"Soft delete product","operationId":"deleteProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/test/verify":{"post":{"tags":["test-controller"],"operationId":"verifyPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordVerifyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/products":{"get":{"tags":["Products"],"summary":"List products","description":"Get paginated list of products","operationId":"listProducts","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}},{"name":"search","in":"query","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"category","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageProductResponse"}}}}}},"post":{"tags":["Products"],"summary":"Create product","description":"Create a new product","operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/products/{productId}/media":{"post":{"tags":["Products"],"summary":"Upload product media","description":"Get presigned URL for uploading product images","operationId":"getMediaUploadUrl","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"fileName","in":"query","required":true,"schema":{"type":"string"}},{"name":"contentType","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediaUploadResponse"}}}}}}},"/api/products/bulk-import":{"post":{"tags":["Products"],"summary":"Bulk import products","description":"Import products from CSV file","operationId":"bulkImport","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BulkImportResponse"}}}}}}},"/api/orders/{orderId}/refund":{"post":{"tags":["Orders"],"summary":"Process refund","description":"Process full or partial refund for an order","operationId":"processRefund","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefundRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RefundResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["Registration"],"summary":"Register new user","description":"Create a new user account","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh access token","description":"Get new access token using refresh token","operationId":"refresh","parameters":[{"name":"X-Refresh-Token","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/auth/logout":{"post":{"tags":["Authentication"],"summary":"User logout","description":"Invalidate current user session","operationId":"logout","responses":{"200":{"description":"OK"}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"User login","description":"Authenticate user with username and password","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/auth/2fa/verify":{"post":{"tags":["Authentication"],"summary":"Verify 2FA code","description":"Verify TOTP code to complete 2FA setup or login","operationId":"verifyTwoFa","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TwoFaVerifyRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/auth/2fa/enable":{"post":{"tags":["Authentication"],"summary":"Enable 2FA","description":"Enable two-factor authentication for current user","operationId":"enableTwoFa","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TwoFaEnableResponse"}}}}}}},"/api/orders/{orderId}/status":{"patch":{"tags":["Orders"],"summary":"Update order status","description":"Update order status with audit trail","operationId":"updateOrderStatus","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDetailResponse"}}}}}}},"/api/test/hash/{password}":{"get":{"tags":["test-controller"],"operationId":"generateHash","parameters":[{"name":"password","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/products/export/csv":{"get":{"tags":["Products"],"summary":"Export products","description":"Export products to CSV","operationId":"exportProducts","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"category","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string","format":"byte"}}}}}}}},"/api/orders":{"get":{"tags":["Orders"],"summary":"List orders","description":"Get paginated list of orders with filters","operationId":"listOrders","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"paymentStatus","in":"query","required":false,"schema":{"type":"string"}},{"name":"fulfillmentStatus","in":"query","required":false,"schema":{"type":"string"}},{"name":"fromDate","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"toDate","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"customerEmail","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageOrderListResponse"}}}}}}},"/api/orders/{orderId}":{"get":{"tags":["Orders"],"summary":"Get order details","description":"Get full order details including items and timeline","operationId":"getOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDetailResponse"}}}}}}},"/api/orders/export/csv":{"get":{"tags":["Orders"],"summary":"Export orders","description":"Export orders to CSV with streaming response","operationId":"exportOrders","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"fromDate","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"toDate","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string","format":"byte"}}}}}}}},"/api/auth/me":{"get":{"tags":["Authentication"],"summary":"Get current user","description":"Retrieve current authenticated user profile","operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}}}}},"/api/auth/2fa/disable":{"delete":{"tags":["Authentication"],"summary":"Disable 2FA","description":"Disable two-factor authentication for current user","operationId":"disableTwoFa","responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"ProductImageDto":{"type":"object","properties":{"url":{"type":"string"},"altText":{"type":"string"}}},"ProductUpdateRequest":{"type":"object","properties":{"name":{"maxLength":500,"minLength":0,"type":"string"},"description":{"type":"string"},"shortDescription":{"maxLength":1000,"minLength":0,"type":"string"},"price":{"minimum":0.0,"exclusiveMinimum":true,"type":"number"},"compareAtPrice":{"minimum":0.0,"exclusiveMinimum":false,"type":"number"},"costPrice":{"minimum":0.0,"exclusiveMinimum":false,"type":"number"},"status":{"type":"string"},"category":{"type":"string"},"brand":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"images":{"type":"array","items":{"$ref":"#/components/schemas/ProductImageDto"}},"weightGrams":{"type":"integer","format":"int32"},"isFeatured":{"type":"boolean"},"seoTitle":{"type":"string"},"seoDescription":{"type":"string"}}},"ProductResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"sku":{"type":"string"},"name":{"type":"string"},"slug":{"type":"string"},"description":{"type":"string"},"shortDescription":{"type":"string"},"price":{"type":"number"},"compareAtPrice":{"type":"number"},"costPrice":{"type":"number"},"status":{"type":"string"},"category":{"type":"string"},"brand":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"images":{"type":"array","items":{"$ref":"#/components/schemas/ProductImageDto"}},"weightGrams":{"type":"integer","format":"int32"},"isFeatured":{"type":"boolean"},"seoTitle":{"type":"string"},"seoDescription":{"type":"string"},"stockQuantity":{"type":"integer","format":"int32"},"lowStock":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"PasswordVerifyRequest":{"type":"object","properties":{"password":{"type":"string"},"hash":{"type":"string"}}},"ProductCreateRequest":{"required":["name","price","sku","slug","status"],"type":"object","properties":{"sku":{"maxLength":100,"minLength":0,"type":"string"},"name":{"maxLength":500,"minLength":0,"type":"string"},"slug":{"maxLength":200,"minLength":0,"pattern":"^[a-z0-9-]+$","type":"string"},"description":{"type":"string"},"shortDescription":{"maxLength":1000,"minLength":0,"type":"string"},"price":{"minimum":0.0,"exclusiveMinimum":true,"type":"number"},"compareAtPrice":{"minimum":0.0,"exclusiveMinimum":false,"type":"number"},"costPrice":{"minimum":0.0,"exclusiveMinimum":false,"type":"number"},"status":{"type":"string"},"category":{"type":"string"},"brand":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"images":{"type":"array","items":{"$ref":"#/components/schemas/ProductImageDto"}},"weightGrams":{"type":"integer","format":"int32"},"isFeatured":{"type":"boolean"},"seoTitle":{"type":"string"},"seoDescription":{"type":"string"}}},"MediaUploadResponse":{"type":"object","properties":{"uploadUrl":{"type":"string"},"fileUrl":{"type":"string"},"key":{"type":"string"}}},"BulkImportResponse":{"type":"object","properties":{"totalRows":{"type":"integer","format":"int32"},"successCount":{"type":"integer","format":"int32"},"failedCount":{"type":"integer","format":"int32"},"errors":{"type":"array","items":{"type":"string"}}}},"RefundRequest":{"required":["amount","reason"],"type":"object","properties":{"amount":{"minimum":0.01,"exclusiveMinimum":false,"type":"number"},"reason":{"type":"string"},"restockItems":{"type":"boolean"}}},"RefundResponse":{"type":"object","properties":{"refundId":{"type":"string","format":"uuid"},"status":{"type":"string"},"amount":{"type":"string"},"message":{"type":"string"}}},"RegisterRequest":{"required":["email","fullName","password","username"],"type":"object","properties":{"username":{"maxLength":50,"minLength":3,"pattern":"^[a-zA-Z0-9_-]+$","type":"string"},"email":{"type":"string"},"password":{"maxLength":100,"minLength":8,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$","type":"string"},"fullName":{"maxLength":100,"minLength":2,"type":"string"}}},"LoginResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserInfo"}}},"UserInfo":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"fullName":{"type":"string"},"roles":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"twoFaEnabled":{"type":"boolean"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"twoFaCode":{"type":"string"}}},"TwoFaVerifyRequest":{"required":["code"],"type":"object","properties":{"code":{"pattern":"\\d{6}","type":"string"}}},"TwoFaEnableResponse":{"type":"object","properties":{"secret":{"type":"string"},"qrCodeUrl":{"type":"string"}}},"OrderStatusUpdateRequest":{"required":["status"],"type":"object","properties":{"status":{"type":"string"},"paymentStatus":{"type":"string"},"fulfillmentStatus":{"type":"string"},"trackingNumber":{"type":"string"},"carrier":{"type":"string"},"notes":{"type":"string"}}},"AddressDto":{"type":"object","properties":{"addressLine1":{"type":"string"},"addressLine2":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"postalCode":{"type":"string"},"country":{"type":"string"}}},"CustomerInfo":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"}}},"OrderDetailResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"orderNumber":{"type":"string"},"customer":{"$ref":"#/components/schemas/CustomerInfo"},"status":{"type":"string"},"paymentStatus":{"type":"string"},"fulfillmentStatus":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemDto"}},"subtotal":{"type":"number"},"taxAmount":{"type":"number"},"shippingAmount":{"type":"number"},"discountAmount":{"type":"number"},"total":{"type":"number"},"couponCode":{"type":"string"},"shippingAddress":{"$ref":"#/components/schemas/AddressDto"},"trackingNumber":{"type":"string"},"carrier":{"type":"string"},"utmSource":{"type":"string"},"utmMedium":{"type":"string"},"utmCampaign":{"type":"string"},"notes":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"OrderItemDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"sku":{"type":"string"},"productName":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"unitPrice":{"type":"number"},"lineTotal":{"type":"number"},"discountAmount":{"type":"number"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageProductResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"},"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"OrderListResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"orderNumber":{"type":"string"},"customerName":{"type":"string"},"customerEmail":{"type":"string"},"status":{"type":"string"},"paymentStatus":{"type":"string"},"fulfillmentStatus":{"type":"string"},"total":{"type":"number"},"itemsCount":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"}}},"PageOrderListResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/OrderListResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"empty":{"type":"boolean"}}},"UserProfileResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"email":{"type":"string"},"fullName":{"type":"string"},"phoneNumber":{"type":"string"},"roles":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"twoFaEnabled":{"type":"boolean"},"lastLoginAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}}}}}